refactor: 카테고리 추천 로직 MenuRecommender -> Categories
